requirements.txt
cv2==1.0
easydict==1.7
numpy==1.12.1
scipy==0.19.0
tensorflow-gpu==1.1.0
Pillow
pyx
